refactor: access token 만료 확인 로직 분리
