--- frontend/src/shared/utils/fetchUtils.ts
@@ -1,3 +1,0 @@
-export const isExpiredAfter60seconds = (tokenExp: number) => {
-  return tokenExp * 1000 - 30 * 1000 < Date.now();
-};
